+++ CHANGES.md
@@ -5,6 +5,10 @@ Change Log
 
 #### next release (8.0.0-alpha.88)
 
+- **Breaking changes**:
+  - `colorPalette` no longer supports a list of CSS colors (eg `rgb(0,0,255)-rgb(0,255,0)-rgb(255,0,0)`). Instead please use `binColors`.
+
+
 * Fixed a bug with numeric item search where it sometimes fails to return all matching values.
 * Respect order of objects from lower strata in `objectArrayTrait`.
 * Fix datetime button margin with scroll in workbench.